python_examples/requirements.txt
pyaml==21.10.1
PyYAML==6.0
pyzmq==24.0.1
zmq==0.0.0
fastapi>=0.95.0
uvicorn>=0.21.0
pydantic>=1.10.7
aiohttp>=3.8.0
python-dotenv>=1.0.0

# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.20.0
httpx>=0.24.0
